wip on component splitting
