+++ src/js/abstract/hocs/with-utils.js
@@ -1,0 +1,9 @@
+export const withUtils = Base =>
+  class extends Base {
+    static uuidv4() {
+      return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+        let r = Math.random() * 16 | 0, v = c == 'x' ? r : (r & 0x3 | 0x8); // eslint-disable-line
+        return v.toString(16);
+      });
+    }
+  };
